separate spherical harmonics and associated legendre
